refactor(section): move section to own folder
