+++ src/data-workspace/data-workspace.js
@@ -20,7 +20,7 @@ const query = {
         resource: 'dataSets',
         id: ({ id }) => id,
         params: {
-            fields: `id,name,displayName,displayFormName,formType,renderAsTabs,
+            fields: `id,name,displayName,displayFormName,formType,renderAsTabs,renderHorizontally,
             dataSetElements[dataElement[id,name,formName,
             categoryCombo[id,name,categories[id,displayName,categoryOptions[id,displayName]],categoryOptionCombos[id,name]]]],
             sections[:owner,displayName,categoryCombos[id,categories[id,displayName,categoryOptions[id,displayName]]]]`,
