+++ src/data-workspace/entry-form.module.css
@@ -1,3 +1,10 @@
+.table {
+    /* overrides fullwidth(100%)- UI-styling */
+    min-width: 240px !important;
+    /* Weird req for allowing <td> cell contents to be 100% cell height */
+    height: 100% !important;
+}
+
 .filterWrapper {
     display: flex;
     align-items: center;
