+++ src/data-workspace/section/section.js
@@ -6,22 +6,19 @@ import {
     TableCellHead,
     TableHead,
     TableRowHead,
-    Tab,
-    TabBar,
 } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useState, useMemo } from 'react'
-import { useSectionFilter } from '../context-selection/use-context-selection/use-context-selection.js'
-import { useMetadata } from '../metadata/index.js'
+import { useMetadata } from '../../metadata/index.js'
 import {
     getDataElementsBySection,
     groupDataElementsByCatCombo,
     groupDataElementsByCatComboInOrder,
-} from '../metadata/selectors.js'
-import { CategoryComboTable } from './category-combo-table.js'
+} from '../../metadata/selectors.js'
+import { CategoryComboTable } from '../category-combo-table.js'
 import styles from './section.module.css'
 
-export const FormSection = ({ section, globalFilterText }) => {
+export const SectionFormSection = ({ section, globalFilterText }) => {
     // Could potentially build table via props instead of rendering children
     const [filterText, setFilterText] = useState('')
     const { available, metadata } = useMetadata()
@@ -115,78 +112,7 @@ const sectionProps = PropTypes.shape({
     id: PropTypes.string,
 })
 
-FormSection.propTypes = {
+SectionFormSection.propTypes = {
     globalFilterText: PropTypes.string,
     section: sectionProps,
-}
-
-export const SectionForms = ({ dataSet, globalFilterText }) => {
-    const [sectionId] = useSectionFilter()
-    const filteredSections = sectionId
-        ? dataSet.sections.filter((s) => s.id === sectionId)
-        : dataSet.sections
-
-    if (dataSet.renderAsTabs) {
-        return (
-            <TabbedSectionForms
-                globalFilterText={globalFilterText}
-                sections={dataSet.sections}
-            />
-        )
-    }
-
-    return (
-        <>
-            {filteredSections.map((s) => (
-                <FormSection
-                    section={s}
-                    key={s.id}
-                    globalFilterText={globalFilterText}
-                />
-            ))}
-        </>
-    )
-}
-
-SectionForms.propTypes = {
-    dataSet: PropTypes.shape({
-        renderAsTabs: PropTypes.bool,
-        sections: PropTypes.arrayOf(sectionProps),
-    }),
-    globalFilterText: PropTypes.string,
-}
-
-const TabbedSectionForms = ({ sections, globalFilterText }) => {
-    const [sectionId, setSelectedId] = useSectionFilter()
-
-    const section = sectionId
-        ? sections.find((s) => s.id === sectionId)
-        : sections[0]
-
-    return (
-        <div>
-            <TabBar className={styles.sectionTab}>
-                {sections.map((s) => (
-                    <Tab
-                        key={s.id}
-                        selected={s.id === section.id}
-                        onClick={() => setSelectedId(s.id)}
-                    >
-                        {s.displayName}
-                    </Tab>
-                ))}
-            </TabBar>
-
-            <FormSection
-                section={section}
-                key={section.id}
-                globalFilterText={globalFilterText}
-            />
-        </div>
-    )
-}
-
-TabbedSectionForms.propTypes = {
-    globalFilterText: PropTypes.string,
-    sections: PropTypes.arrayOf(sectionProps),
 }
